product_service/requirements.txt
fastapi
uvicorn
pika
grpcio
grpcio-tools
sqlalchemy
psycopg2-binary